Adding missing welding mask sprites for Feroxi and Vulpkanin (#3548)
* weld gas mask and normal gas mask additions

idk if this is the correct way to do it but it works. I didn't need to touch the yml for masks.yml to get them to work on vulp and feroxi. Toggles work for em and the sprites load. ⭐

* Welding masks not fit the shorks

their goofy ahh snouts don't stick out now and they now follow proper PPE protocols

* Vulpkanin more like vulpkanin

Touch up on copyright

Signed-off-by: Zephyr <68482449+Zephyrpak@users.noreply.github.com>

---------

Signed-off-by: Zephyr <68482449+Zephyrpak@users.noreply.github.com>